DateFix patch: this ensures that the built NSIS installer behaves the same way as regular zip (eg 7zip etc) when packing files up in one timezone and installing those files (ie unpacking) in a different timezone. Without this "DF" (DateFix) patch, the unpacked files will change their dates due to the timezone shift.
Compare this to the behaviour of 7zip (or other standard zip packages). If you zip files in one timezone, then download/transfer the
zip to another timezone where you unpack it, the directory listing of the unpacked files will show the same dates and times as the
original directory tree that was zipped. We want NSIS installer EXEs built to also preserve these date/time stamps, as in some
applications, timestamps of various files are compared to see if things are up to date, etc.

This DF (DateFix) patch fixes this problem.
We have been using this on all our NSIS installers for more than a decade, but I believe
it belongs in the standard src tree. Is there any reason to behave differently to major zip utilities like 7zip?